requirements.txt
#Framework packages
django
djangorestframework
drf-spectacular

#Dynamodb ORM packages
pynamodb